+++ examples/simultaneous-cursors/src/App.jsx
@@ -8,7 +8,9 @@ const client = new yorkie.Client(import.meta.env.VITE_YORKIE_API_ADDR, {
   apiKey: import.meta.env.VITE_YORKIE_API_KEY,
 });
 
-const doc = new yorkie.Document('simultaneous-cursors');
+const doc = new yorkie.Document('simultaneous-cursors', {
+  enableDevtools: true,
+});
 
 const App = () => {
   const [clients, setClients] = useState([]);
@@ -87,6 +89,7 @@ const App = () => {
     <div className="general-container">
       {clients.map(
         ({ clientID, presence: { cursorShape, cursor, pointerDown } }) => {
+          if (!cursor) return null;
           return (
             <Cursor
               selectedCursorShape={cursorShape}