+++ examples/vanilla-quill/src/type.ts
@@ -6,5 +6,6 @@ export type YorkieDoc = {
 
 export type YorkiePresence = {
   username: string;
+  color: string;
   selection: TextPosStructRange | undefined;
 };