+++ src/devtools/types.ts
@@ -15,6 +15,7 @@
  */
 
 import type { PrimitiveValue } from '@yorkie-js-sdk/src/document/crdt/primitive';
+import type { CRDTTreePosStruct } from '@yorkie-js-sdk/src/document/crdt/tree';
 import { CounterValue } from '@yorkie-js-sdk/src/document/crdt/counter';
 
 /**
@@ -78,7 +79,7 @@ export type ContainerValue = {
 };
 
 /**
- * `TreeNodeInfo` represents the tree node information in devtools.
+ * `TreeNodeInfo` represents the crdt tree node information in devtools.
  */
 export type TreeNodeInfo = {
   id: string;
@@ -91,5 +92,9 @@ export type TreeNodeInfo = {
   insPrev?: string;
   insNext?: string;
   children: Array<TreeNodeInfo>;
+  attributes?: object; // TODO(chacha912): Specify the type accurately.
   depth: number;
+  index?: number;
+  path?: Array<number>;
+  pos?: CRDTTreePosStruct;
 };