+++ src/document/time/ticket.ts
@@ -50,7 +50,6 @@ export class TimeTicket {
   private delimiter: number;
   private actorID: ActorID;
 
-  /** @hideconstructor */
   constructor(lamport: Long, delimiter: number, actorID: string) {
     this.lamport = lamport;
     this.delimiter = delimiter;