+++ src/yorkie.ts
@@ -20,45 +20,47 @@ import { Text } from '@yorkie-js-sdk/src/document/json/text';
 import { Tree } from '@yorkie-js-sdk/src/document/json/tree';
 import { Counter } from '@yorkie-js-sdk/src/document/json/counter';
 import { CounterType } from '@yorkie-js-sdk/src/document/crdt/counter';
-import type * as Devtools from '@yorkie-js-sdk/src/devtools/types';
+import * as Devtools from '@yorkie-js-sdk/src/devtools/types';
 
 export {
   Client,
-  ClientEvent,
   ClientStatus,
   SyncMode,
-  StreamConnectionStatus,
-  DocumentSyncResultType,
-  ClientEventType,
-  StatusChangedEvent,
-  DocumentChangedEvent,
-  StreamConnectionStatusChangedEvent,
-  DocumentSyncedEvent,
-  ClientOptions,
+  type ClientOptions,
 } from '@yorkie-js-sdk/src/client/client';
 export {
   DocEventType,
-  SnapshotEvent,
-  LocalChangeEvent,
-  RemoteChangeEvent,
-  Indexable,
-  DocEvent,
+  type SnapshotEvent,
+  type LocalChangeEvent,
+  type RemoteChangeEvent,
+  type ConnectionChangedEvent,
+  type SyncStatusChangedEvent,
+  type WatchedEvent,
+  type UnwatchedEvent,
+  type PresenceChangedEvent,
+  type InitializedEvent,
+  StreamConnectionStatus,
+  DocumentSyncStatus,
+  DocumentStatus,
+  type Indexable,
+  type DocEvent,
+  type TransactionEvent,
   Document,
-  ChangeInfo,
+  type ChangeInfo,
 } from '@yorkie-js-sdk/src/document/document';
 export {
-  Observer,
-  Observable,
-  NextFn,
-  ErrorFn,
-  CompleteFn,
-  Unsubscribe,
+  type Observer,
+  type Observable,
+  type NextFn,
+  type ErrorFn,
+  type CompleteFn,
+  type Unsubscribe,
 } from '@yorkie-js-sdk/src/util/observable';
 export {
   TimeTicket,
-  TimeTicketStruct,
+  type TimeTicketStruct,
 } from '@yorkie-js-sdk/src/document/time/ticket';
-export { ActorID } from '@yorkie-js-sdk/src/document/time/actor_id';
+export { type ActorID } from '@yorkie-js-sdk/src/document/time/actor_id';
 export type {
   OperationInfo,
   TextOperationInfo,
@@ -74,35 +76,36 @@ export type {
   EditOpInfo,
   StyleOpInfo,
 } from '@yorkie-js-sdk/src/document/operation/operation';
+export { OpSource } from '@yorkie-js-sdk/src/document/operation/operation';
 
 export {
   Primitive,
-  PrimitiveValue,
+  type PrimitiveValue,
 } from '@yorkie-js-sdk/src/document/crdt/primitive';
 import { Primitive } from '@yorkie-js-sdk/src/document/crdt/primitive';
 
 export {
-  WrappedElement,
-  JSONElement,
+  type WrappedElement,
+  type JSONElement,
 } from '@yorkie-js-sdk/src/document/json/element';
-export { JSONObject } from '@yorkie-js-sdk/src/document/json/object';
-export { JSONArray } from '@yorkie-js-sdk/src/document/json/array';
+export { type JSONObject } from '@yorkie-js-sdk/src/document/json/object';
+export { type JSONArray } from '@yorkie-js-sdk/src/document/json/array';
 export { Counter } from '@yorkie-js-sdk/src/document/json/counter';
-export { CounterValue } from '@yorkie-js-sdk/src/document/crdt/counter';
+export { type CounterValue } from '@yorkie-js-sdk/src/document/crdt/counter';
 export {
   Text,
-  TextPosStruct,
-  TextPosStructRange,
+  type TextPosStruct,
+  type TextPosStructRange,
 } from '@yorkie-js-sdk/src/document/json/text';
 export {
   Tree,
-  TreeNode,
-  ElementNode,
-  TextNode,
-  TreeChange,
-  TreeChangeType,
-  CRDTTreeNodeIDStruct,
-  TreePosStructRange,
+  type TreeNode,
+  type ElementNode,
+  type TextNode,
+  type TreeChange,
+  type TreeChangeType,
+  type CRDTTreeNodeIDStruct,
+  type TreePosStructRange,
 } from '@yorkie-js-sdk/src/document/json/tree';
 export { Change } from '@yorkie-js-sdk/src/document/change/change';
 export { converter } from '@yorkie-js-sdk/src/api/converter';
@@ -113,12 +116,12 @@ export { setLogLevel } from '@yorkie-js-sdk/src/util/logger';
 export {
   EventSourceDevPanel,
   EventSourceSDK,
-  PanelToSDKMessage,
-  SDKToPanelMessage,
-  FullPanelToSDKMessage,
-  FullSDKToPanelMessage,
+  type PanelToSDKMessage,
+  type SDKToPanelMessage,
+  type FullPanelToSDKMessage,
+  type FullSDKToPanelMessage,
 } from '@yorkie-js-sdk/src/devtools/protocol';
-export type { Devtools };
+export { Devtools };
 
 /**
  * The top-level yorkie namespace with additional properties.
@@ -130,7 +133,7 @@ export type { Devtools };
  *
  * @public
  */
-const yorkie = {
+export default {
   Client,
   Document,
   Primitive,
@@ -141,4 +144,16 @@ const yorkie = {
   LongType: CounterType.LongCnt,
 };
 
-export default yorkie;
+// TODO(hackerwins): Remove this when we have a better way to expose the API.
+if (typeof globalThis !== 'undefined') {
+  (globalThis as any).yorkie = {
+    Client,
+    Document,
+    Primitive,
+    Text,
+    Counter,
+    Tree,
+    IntType: CounterType.IntegerCnt,
+    LongType: CounterType.LongCnt,
+  };
+}