+++ test/integration/tree_test.ts
@@ -15,7 +15,7 @@
  */
 
 import { describe, it, assert } from 'vitest';
-import yorkie, { Tree } from '@yorkie-js-sdk/src/yorkie';
+import yorkie, { Tree, SyncMode } from '@yorkie-js-sdk/src/yorkie';
 import {
   testRPCAddr,
   toDocKey,
@@ -1335,6 +1335,42 @@ describe('Tree.style', function () {
     });
   });
 
+  it('Can sync its content with remove style', async function ({ task }) {
+    await withTwoClientsAndDocuments<{ t: Tree }>(async (c1, d1, c2, d2) => {
+      d1.update((root) => {
+        root.t = new Tree({
+          type: 'doc',
+          children: [
+            {
+              type: 'p',
+              children: [{ type: 'text', value: 'hello' }],
+              attributes: { italic: 'true' },
+            },
+          ],
+        });
+      });
+      await c1.sync();
+      await c2.sync();
+      assert.equal(
+        d1.getRoot().t.toXML(),
+        /*html*/ `<doc><p italic="true">hello</p></doc>`,
+      );
+      assert.equal(
+        d2.getRoot().t.toXML(),
+        /*html*/ `<doc><p italic="true">hello</p></doc>`,
+      );
+
+      d1.update((root) => {
+        root.t.removeStyle(0, 1, ['italic']);
+      });
+      await c1.sync();
+      await c2.sync();
+
+      assert.equal(d1.getRoot().t.toXML(), /*html*/ `<doc><p>hello</p></doc>`);
+      assert.equal(d2.getRoot().t.toXML(), /*html*/ `<doc><p>hello</p></doc>`);
+    }, task.name);
+  });
+
   it('Should return correct range path within doc.subscribe', async function ({
     task,
   }) {
@@ -1496,8 +1532,8 @@ describe('Tree.style', function () {
     await c1.activate();
     await c2.activate();
 
-    await c1.attach(d1, { isRealtimeSync: false });
-    await c2.attach(d2, { isRealtimeSync: false });
+    await c1.attach(d1, { syncMode: SyncMode.Manual });
+    await c2.attach(d2, { syncMode: SyncMode.Manual });
 
     // Perform a dummy update to apply changes up to the snapshot threshold.
     const snapshotThreshold = 500;
@@ -1697,7 +1733,7 @@ describe('Tree.style', function () {
     const d3 = new yorkie.Document<TestDoc>(docKey);
     const c3 = new yorkie.Client(testRPCAddr);
     await c3.activate();
-    await c3.attach(d3, { isRealtimeSync: false });
+    await c3.attach(d3, { syncMode: SyncMode.Manual });
     assert.equal(
       d3.getRoot().t.toXML(),
       /*html*/ `<r><c><u><p><n></n></p></u></c><c><p><n>1 카카오2 네이3</n></p></c></r>`,
@@ -4364,14 +4400,14 @@ describe('TreeChange', () => {
           } as any,
           {
             type: 'tree-edit',
-            from: 1,
-            to: 2,
+            from: 3,
+            to: 4,
             value: undefined,
           } as any,
           {
             type: 'tree-edit',
-            from: 3,
-            to: 4,
+            from: 1,
+            to: 2,
             value: undefined,
           } as any,
         ],